10608: Changes as per PR comments.
